Adds the compute node end of the logging pipeline
Adds a new endpoint in the compute node to allow a caller to request a
multiaddr where the logserver is listening for new connections (using a
protocol ID of its own). Logging was not added to bprotocol directly as
it is a request/response protocol, whereas logging is a streaming,
long-lived response.