Handle completed jobs
When running a job, sometimes it errors and if we are following the logs
we get no output (as we cannot get the streams when the job has finished). To handle this, if we determine that the job/execution has already completed (successfully or not) then we will return the job output on the appropriate stream.